Update index.rst
Salt itself now contains salt-cloud. As these instructions are written, they do not work because when you clone salt-cloud, the main source is not in this top-level directory but on level lower in the same-named directory (salt-cloud). 

The README in the salt-cloud docs states: 

"The Salt team is excited to report that the ``salt-cloud`` repository has been
merged into the main Salt repository."